+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Expect active development and potentially significant breaking changes in the `0
 
 ### vNEXT
 
+- Catch uncaught promise errors in QueryManager [#950](https://github.com/apollostack/apollo-client/pull/950)
+
 ### 0.5.6
 - Refactor polling query logic to fix startPolling and stopPolling [#938](https://github.com/apollostack/apollo-client/pull/938)
 - Add convenience method to obtain initial state from SSR [#941](https://github.com/apollostack/apollo-client/pull/941)